Create mixin: movingTasksAndColumnsMixin for components
because these components are sharing duplicate data and methods
