+++ Practical-Builds/clone-trello/src/components/ColumnTask.vue
@@ -22,7 +22,10 @@
 </template>
 
 <script>
+import movingTasksAndColumnsMixin from '@/mixins/movingTasksAndColumnsMixin'
+
 export default {
+    mixins: [movingTasksAndColumnsMixin],
     props: {
         task: {
             type: Object,
@@ -30,18 +33,6 @@ export default {
         },
         taskIndex: {
             type: Number,
-            required: true
-        },
-        column: {
-            type: Object,
-            required: true
-        },
-        columnIndex: {
-            type: Number,
-            required: true
-        },
-        board: {
-            type: Object,
             required: true
         }
     },
@@ -55,33 +46,6 @@ export default {
             e.dataTransfer.setData('from-task-index', taskIndex)
             e.dataTransfer.setData('from-column-index', fromColumnIndex)
             e.dataTransfer.setData('type', 'task')
-        },
-        moveTaskOrColumn (e, toTasks, toColumnIndex, toTaskIndex) {
-            const type = e.dataTransfer.getData('type')
-            if (type === 'task') {
-                this.moveTask(e, toTasks)
-            } else {
-                this.moveColumn(e, toColumnIndex)
-            }
-        },
-        moveTask (e, toTasks, toTaskIndex) {
-            const fromColumnIndex = e.dataTransfer.getData('from-column-index')
-            const fromTasks = this.board.columns[fromColumnIndex].tasks
-            const fromTaskIndex = e.dataTransfer.getData('from-task-index')
-
-            this.$store.commit('MOVE_TASK', {
-                fromTasks,
-                fromTaskIndex,
-                toTasks,
-                toTaskIndex
-            })
-        },
-        moveColumn (e, toColumnIndex) {
-            const fromColumnIndex = e.dataTransfer.getData('from-column-index')
-            this.$store.commit('MOVE_COLUMN', {
-                fromColumnIndex,
-                toColumnIndex
-            })
         }
     }
 }
