+++ Practical-Builds/clone-trello/src/mixins/movingTasksAndColumnsMixin.js
@@ -1,0 +1,45 @@
+export default {
+    props: {
+        column: {
+            type: Object,
+            required: true
+        },
+        columnIndex: {
+            type: Number,
+            required: true
+        },
+        board: {
+            type: Object,
+            required: true
+        }
+    },
+    methods: {
+        moveTaskOrColumn (e, toTasks, toColumnIndex, toTaskIndex) {
+            const type = e.dataTransfer.getData('type')
+            if (type === 'task') {
+                this.moveTask(e, toTasks)
+            } else {
+                this.moveColumn(e, toColumnIndex)
+            }
+        },
+        moveTask (e, toTasks, toTaskIndex) {
+            const fromColumnIndex = e.dataTransfer.getData('from-column-index')
+            const fromTasks = this.board.columns[fromColumnIndex].tasks
+            const fromTaskIndex = e.dataTransfer.getData('from-task-index')
+
+            this.$store.commit('MOVE_TASK', {
+                fromTasks,
+                fromTaskIndex,
+                toTasks,
+                toTaskIndex
+            })
+        },
+        moveColumn (e, toColumnIndex) {
+            const fromColumnIndex = e.dataTransfer.getData('from-column-index')
+            this.$store.commit('MOVE_COLUMN', {
+                fromColumnIndex,
+                toColumnIndex
+            })
+        }
+    }
+}
